restructure code, remove cruft
